Behaviour tree
Implements a behaviour tree to predict the narrator actions based
on the player interactions with him. For now, the player interacts
with the narrator by choosing to press or not the space key.

Each interaction can result on different endings. The information
about the choices and behaviours is loaded from a xml file.