First set of rooms with furnishings
Working in "Purple scene"; The first set of rooms with models without
textures blocked into them. The vase and mirror are currently
non-breakable.

The walls have a lower section to remain when the main section is
excluded via the raycast. The basic room layout is grouped under "Base
room" to enable easy duplication for more rooms or experimenting.

There is also a currently in-progress particle effect test for the room
where Aurora Australis is to appear as part of the first "Happy ending".