Upgraded behaviour tree mechanism
Upgrades and extends the behaviour tree mechanism so it can be
more flexible and can be dowloaded from a twee file (that is
generated from a twine project).

For now there is a known bug when it comes to linking the nodes.
During tests it can be seem that the right node is being correctly
linked but the left node is not.