Behaviour Tree that uses commands parsed from twee
Implements methods on the BehaviourTree script that are used to
communicate with other scripts. Thus, now is possible to use
information collected from the twee file for the mechanism that
trigger events involving other objects.

Also, adds a "actionbutton" command that allows player to
communicate with narrator through "spacebar" key.

For now the tests showed that the event triggered by time and the
event triggered by action button are working correctly. However,
there are still known bugs when it comes to trigger an event by
interacting with a specific object.