Narrator controller
Add the narrator.
For using place the audio file on: Assets/Resources/Narrator/ and the
subtitles on the Assets/Resources/Subtitles/, both with the same name (but
different extensions, as expected)

For starting, execute the public method playDialog with the name of the
file.

Subtitle should be: time(in seconds)|text(or \clear to remove the text)
example:
00.00|text that the subtitle should display
01.97|\clear

In the subtitle is also possible to call other objects methods synced with
the audio using \call-(ObjectName)-(MethodName)[-(string parameter)]
01.50|\call-1STARTROOM/Door 1-Open