Merging Tree with purple scene and subtitles
The twine project was updated to use the syntax that the twee
parser uses, to create a behaviour tree compactible with the game
story.

Now some of the elements of the purple scene are used to make
choice triggers.

Also, the Behaviour Tree now uses the NarratorController to set
audio and subtitles syncronized with the player choices.