Rooms different colours, new models, etc
The rooms have had their colours changed to look more alike to the
original concept art, and to look more cohesive. Materials used on items
within rooms are grouped within the appropriately named folder. For
rooms that require a palette shift during the course of the game,
grouped within the same folder is the alternative colour on a material
that just has "ALT" as a prefix to the name of the object it is to be
used on. There are also a few new decorative models and textures on some
items.

The rooms have also been changed to have their floor align cleaner
between the door and the next room. Some rooms have been made slightly
larger, to be adjusted as we figure out more appropriate sizes for each
individual room. In addition, all rooms now have higher walls in
addition to the camera being closer to the character so that the player
cannot view the skybox/next room without passing through the door.