Animations, models
Doll model now in scene along with a collider. The edges on the ladies
chair in room 3 have also been softened.

The animations for both the mirror shattering and the vase toppling and
breaking are now in. Currently they play on awake, must be delayed to
only animate when player comes within range.