Adding the music triggers
Now to call a music trigger can be done using the MusicController on the
subtitle file like this:

time |\call-Abstract/MusicController-PlayAudio-AUDIO_NAME
00.67|\call-Abstract/MusicController-PlayAudio-happy

The "happy" and "sad" are special cases, e.g. happy will one file from the
folder Happy sequencially between calls.

And for calling an specific file, just use the name of the file on
Assets/Resources/Soundtrack